Allow for custom search.
The built-in search only allows for matching on cluster name.  This
allows a custom search query to be provided.

My use case is I may have a cluster_name that is a chef environment,
but I only want nodes in that env with a particular role to match
the search (eg chef_environment:dev AND role:database).

If you dont specify a custom_search_query, you fall through to the old
static search.