Prevent removal of MariaDB-compat for centos-7 #2
When mariadb is installed with a package at a version different than
the default os one, during removal of mysql-libs with yum multiple
matches can be found but chef could only use the first match:
MariaDB-compat. Therefore, because of dependencies, the removal of
MariaDB-compat triggers the one of MariaDB-server and several other
packages (MariaDB-shared, MariaDB-client, MariaDB-common, MariaDB-devel,
MySQL-python). To fix this issue we prevent the removal of mysql-libs
for centos-7 and later.

Signed-off-by: Jean-Francois Weber-Marx <jfwm@hotmail.com>